+++ struckoutCameraApp/app/src/main/java/com/taichi765/struckoutCameraApp/camera/types/CameraLocation.kt
@@ -1,3 +1,3 @@
 package com.taichi765.struckoutCameraApp.camera.types
 
-data class CameraLocation(val x: Float, val y: Float, val z: Float)
+data class CameraLocation(val x: Double, val y: Double, val z: Double)
